🐛 FIX: Fixed neovim and insync; added absolute paths for initial setup

+++ scripts/setup.sh
@@ -1,6 +1,15 @@
 #!/bin/bash
 
-dotfiles_temp=$(echo `pwd` | sed "s/\/scripts//")
+DOTFILES=$HOME/.dotfiles
+
+read -p "The path of dotfiles is set at $DOTFILES . Is this correct? [Y]es [n]o: " answer
+answer=$(echo "$answer" | awk '{print tolower($0)}')
+
+if [[ $answer == "n" || $answer == "no" ]]
+then
+	read -p "What is the correct path? " path
+	DOTFILES=$path
+fi
 
 # SETTINGS
 ## Brightness without sudo
@@ -9,8 +18,12 @@ sudo chmod 4775 /usr/bin/brightnessctl
 
 # SYMLINKS
 ## bashrc
-mv .bashrc .bashrc_old
-ln -s $dotfiles_temp/.bashrc $HOME
+if [ -f $HOME/.bashrc ]
+then
+	mv .bashrc .bashrc_old
+fi
+
+ln -s $DOTFILES/.bashrc --target-directory=$HOME
 source $HOME/.bashrc
 
 ## .config files
@@ -40,3 +53,8 @@ sudo ln -s $DOTFILES/xorg.conf.d/30-touchpad.conf /usr/share/X11/xorg.conf.d
 
 ## eww config
 bash $SCRIPTS/eww.sh
+
+echo
+echo "Setup done."
+echo "Go to $HOME/.bashrc and check that all exported paths are correct."
+exit 0
